move near config to config file
