+++ aurora/configs/aurora-mainnet.json
@@ -8,5 +8,11 @@
     "auroraUtilsAddress": "0xF59eFA6A6065674D639d8b5f1FDdADf7F8E5E8E2",
     "auroraCodecAddress": "0xC042D3F70ADc2f74a1717147f9DDF413fee487C1",
     "wNearAddress": "0xc42c30ac6cc15fac9bd938618bcaa1a1fae8501d",
-    "nearNetwork": "mainnet"
+    "nearConnectionConfig": {
+        "networkId": "mainnet",
+        "nodeUrl": "https://rpc.mainnet.near.org",
+        "walletUrl": "https://wallet.mainnet.near.org",
+        "helperUrl": "https://helper.mainnet.near.org",
+        "explorerUrl": "https://explorer.mainnet.near.org"
+    }
 }
